+++ R/string.R
@@ -1,18 +1,16 @@
 #' Test if a character vector consists of blank strings
 #'
-#' Return `TRUE` if all elements of a character vector are blank (white
-#' spaces or empty strings).
+#' Return a logical vector indicating if elements of a character vector are
+#' blank (white spaces or empty strings).
 #' @param x A character vector.
-#' @return `TRUE` if the input is blank, or `FALSE` otherwise.
+#' @return `TRUE` for blank elements, or `FALSE` otherwise.
 #' @export
 #' @examples
 #' xfun::is_blank('')
 #' xfun::is_blank('abc')
 #' xfun::is_blank(c('', '  ', '\n\t'))
 #' xfun::is_blank(c('', ' ', 'abc'))
-is_blank = function(x) {
-  all(grepl('^\\s*$', x))
-}
+is_blank = function(x) grepl('^\\s*$', x)
 
 #' Convert numbers to English words
 #'